backported VBE and 3dgfx fixes from dosdemo, dropped MMX fade

+++ src/sprite.c
@@ -194,7 +194,7 @@ void draw_sprite(void *dest, int fbpitch, struct sprites *ss, int idx)
 			xoffs += sop->size;
 			break;
 		case SOP_FILL:
-			memset16(fbptr + xoffs, (uint16_t)sop->data, sop->size >> 1);
+			memset16(fbptr + xoffs, (uintptr_t)sop->data, sop->size >> 1);
 			xoffs += sop->size;
 			break;
 		default:
